+++ wemake_python_styleguide/violations/consistency.py
@@ -43,7 +43,9 @@ Summary
    BadNumberSuffixViolation
    ComparisonOrderViolation
    MultipleInComparisonViolation
+   RedundantComparisonViolation
    NamedConstantConditionalViolation
+
 
 Consistency checks
 ------------------
@@ -60,6 +62,7 @@ Consistency checks
 .. autoclass:: BadNumberSuffixViolation
 .. autoclass:: ComparisonOrderViolation
 .. autoclass:: MultipleInComparisonViolation
+.. autoclass:: RedundantComparisonViolation
 .. autoclass:: NamedConstantConditionalViolation
 
 """
@@ -433,11 +436,10 @@ class BadNumberSuffixViolation(TokenizeViolation):
 
 class MultipleInComparisonViolation(ASTViolation):
     """
-    Forbids comparision where multiple 'in's are userd in a statement.
+    Forbids comparision where multiple ``in`` checks.
 
     Reasoning:
-        This is unreadable. Use different comparisons for it.
-        Bring a consistency to the comparison!
+        Using multiple ``in`` is unreadable.
 
     Solution:
         Refactor your comparison expression to use several ``and`` conditions
@@ -460,8 +462,42 @@ class MultipleInComparisonViolation(ASTViolation):
 
     should_use_text = False
     #: Error message shown to the user.
-    error_template = 'Found multiple in comparisons'
+    error_template = 'Found multiple `in` comparisons'
     code = 311
+
+
+class RedundantComparisonViolation(ASTViolation):
+    """
+    Forbids to have comparisons between the same variable.
+
+    Reasoning:
+        When the same variables are compared it is typically an indication
+        of a mistake, since the Boolean value of the comparison will always be
+        the same.
+
+    Solution:
+        Remove the same variable comparison and any associated dead code.
+
+    Example::
+
+        # Wrong:
+        a = 1
+        if a < a:
+            do_something()
+        else:
+            do_something_else()
+
+        # Correct:
+        do_something()
+
+    Note:
+        Returns Z312 as error code
+    """
+
+    should_use_text = False
+    #: Error message shown to the user.
+    error_template = 'Found comparison between same variable'
+    code = 312
 
 
 class NamedConstantConditionalViolation(ASTViolation):
@@ -486,7 +522,6 @@ class NamedConstantConditionalViolation(ASTViolation):
 
     Note:
         Returns Z313 as error code
-
     """
 
     should_use_text = False