+++ wemake_python_styleguide/visitors/ast/naming.py
@@ -1,17 +1,13 @@
 # -*- coding: utf-8 -*-
 
 import ast
+from typing import Union
 
 from wemake_python_styleguide.constants import (
     MODULE_METADATA_VARIABLES_BLACKLIST,
     VARIABLE_NAMES_BLACKLIST,
 )
-from wemake_python_styleguide.logics.variables import (
-    is_private_variable,
-    is_too_short_variable_name,
-    is_variable_name_with_underscored_number,
-    is_wrong_variable_name,
-)
+from wemake_python_styleguide.logics import variables
 from wemake_python_styleguide.types import AnyFunctionDef, AnyImport
 from wemake_python_styleguide.violations.best_practices import (
     WrongModuleMetadataViolation,
@@ -20,10 +16,13 @@ from wemake_python_styleguide.violations.naming import (
     PrivateNameViolation,
     TooShortVariableNameViolation,
     UnderScoredNumberNameViolation,
+    UpperCaseAttributeViolation,
     WrongVariableNameViolation,
 )
 from wemake_python_styleguide.visitors.base import BaseNodeVisitor
 from wemake_python_styleguide.visitors.decorators import alias
+
+VariableDef = Union[ast.Name, ast.Attribute, ast.ExceptHandler]
 
 
 @alias('visit_any_import', (
@@ -34,26 +33,26 @@ from wemake_python_styleguide.visitors.decorators import alias
     'visit_FunctionDef',
     'visit_AsyncFunctionDef',
 ))
+@alias('visit_variable', (
+    'visit_Name',
+    'visit_Attribute',
+    'visit_ExceptHandler',
+))
 class WrongNameVisitor(BaseNodeVisitor):
-    """
-    Performs checks based on variable names.
-
-    It is responsible for finding short and blacklisted variables,
-    functions, and arguments.
-    """
+    """Performs checks based on variable names."""
 
     def _check_name(self, node: ast.AST, name: str) -> None:
-        if is_wrong_variable_name(name, VARIABLE_NAMES_BLACKLIST):
+        if variables.is_wrong_variable_name(name, VARIABLE_NAMES_BLACKLIST):
             self.add_violation(WrongVariableNameViolation(node, text=name))
 
         min_length = self.options.min_variable_length
-        if is_too_short_variable_name(name, min_length=min_length):
+        if variables.is_too_short_variable_name(name, min_length=min_length):
             self.add_violation(TooShortVariableNameViolation(node, text=name))
 
-        if is_private_variable(name):
+        if variables.is_private_variable(name):
             self.add_violation(PrivateNameViolation(node, text=name))
 
-        if is_variable_name_with_underscored_number(name):
+        if variables.is_variable_name_with_underscored_number(name):
             self.add_violation(UnderScoredNumberNameViolation())
 
     def _check_function_signature(self, node: AnyFunctionDef) -> None:
@@ -69,19 +68,30 @@ class WrongNameVisitor(BaseNodeVisitor):
         if node.args.kwarg:
             self._check_name(node, node.args.kwarg.arg)
 
-    def visit_Attribute(self, node: ast.Attribute) -> None:
+    def _check_attribute_name(self, node: ast.ClassDef) -> None:
+        top_level_assigns = [
+            sub_node
+            for sub_node in node.body
+            if isinstance(sub_node, ast.Assign)
+        ]
+
+        for assignment in top_level_assigns:
+            for target in assignment.targets:
+                name = getattr(target, 'id', None)
+                if variables.is_upper_case_name(name):
+                    self.add_violation(
+                        UpperCaseAttributeViolation(target, text=name),
+                    )
+
+    def visit_ClassDef(self, node: ast.ClassDef) -> None:
         """
-        Used to find wrong attribute names inside classes.
+        Used to find upper attribute declarations.
 
         Raises:
-            WrongVariableNameViolation
-            TooShortVariableNameViolation
-            PrivateNameViolation
+            UpperCaseAttributeViolation
 
         """
-        if isinstance(node.ctx, ast.Store):
-            self._check_name(node, node.attr)
-
+        self._check_attribute_name(node)
         self.generic_visit(node)
 
     def visit_any_function(self, node: AnyFunctionDef) -> None:
@@ -98,34 +108,6 @@ class WrongNameVisitor(BaseNodeVisitor):
         self._check_function_signature(node)
         self.generic_visit(node)
 
-    def visit_ExceptHandler(self, node: ast.ExceptHandler) -> None:
-        """
-        Used to find wrong exception instances in ``try``/``except``.
-
-        Raises:
-            WrongVariableNameViolation
-            TooShortVariableNameViolation
-            PrivateNameViolation
-
-        """
-        self._check_name(node, getattr(node, 'name', None))
-        self.generic_visit(node)
-
-    def visit_Name(self, node: ast.Name) -> None:
-        """
-        Used to find wrong regular variables.
-
-        Raises:
-            WrongVariableNameViolation
-            TooShortVariableNameViolation
-            PrivateNameViolation
-
-        """
-        if isinstance(node.ctx, ast.Store):
-            self._check_name(node, node.id)
-
-        self.generic_visit(node)
-
     def visit_any_import(self, node: AnyImport) -> None:
         """
         Used to check wrong import alias names.
@@ -140,6 +122,21 @@ class WrongNameVisitor(BaseNodeVisitor):
             if alias_node.asname:
                 self._check_name(node, alias_node.asname)
 
+        self.generic_visit(node)
+
+    def visit_variable(self, node: VariableDef) -> None:
+        """
+        Used to check wrong names of assigned variables.
+
+        Raises:
+            WrongVariableNameViolation
+            TooShortVariableNameViolation
+            PrivateNameViolation
+
+        """
+        variable_name = variables.get_assigned_name(node)
+        if variable_name is not None:
+            self._check_name(node, variable_name)
         self.generic_visit(node)